fix: clear listeners after sending destroy event, not before
If listeners are cleared before sending the 'destroy' event, the event
is useless, because nobody is listening for it anymore.

This change has the side effect of also sending a 'remove' event for
each child when a parent is destroyed, since that had also been done
after clearing listeners but before sending the 'destroy' event.